switched u32 to usize in multiple structs/functions. also found a way to insert unique hash key by using len of hashmap

+++ src/geography.rs
@@ -2,36 +2,36 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct World {
-    id: u32,
+    id: usize,
     name: String,
 }
 
 impl World {
-    pub fn new(id: u32, name: String) -> Self {
+    pub fn new(id: usize, name: String) -> Self {
         World { id, name }
     }
 }
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Continent {
-    pub id: u32,
+    pub id: usize,
     pub name: String,
 }
 
 impl Continent {
-    pub fn new(id: u32, name: String) -> Self {
+    pub fn new(id: usize, name: String) -> Self {
         Continent { id, name }
     }
 }
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Region {
-    pub id: u32,
+    pub id: usize,
     pub name: String,
 }
 
 impl Region {
-    pub fn new(id: u32, name: String) -> Self {
+    pub fn new(id: usize, name: String) -> Self {
         Region { id, name }
     }
 }
@@ -39,9 +39,9 @@ impl Region {
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Province {
     // we live in a province; bottom text
-    id: u32,
+    id: usize,
     name: String,
-    resource_id: u32,
+    resource_id: usize,
     //is_sea: bool,
     //is_coastal: bool,
     //is_lake: bool,
@@ -54,7 +54,7 @@ pub struct Province {
 }
 
 impl Province {
-    pub fn new(id: u32, name: String, resource_id: u32) -> Province {
+    pub fn new(id: usize, name: String, resource_id: usize) -> Province {
         Province {
             id,
             name,
